Extract a AddressGeocoder service and add support for explicitly requesting an address to be geocoded on the Api::EnterprisesController too.
The new feature which allows people to select their location on a map during signup only works with Google maps so far. So if an instance is using Open Street Map this change will allow us to pass a :use_geocoder paramter to the Api::EnterprisesController so that the address will be geocoded on the backend without the use of a map.